refactor: 설정 화면 관련 모델 패키지로 이동
