+++ android/app/src/main/java/com/mulkkam/ui/setting/SettingScreen.kt
@@ -19,6 +19,7 @@ import com.mulkkam.ui.designsystem.MulkkamTheme
 import com.mulkkam.ui.setting.component.SettingDividerItem
 import com.mulkkam.ui.setting.component.SettingNormalItem
 import com.mulkkam.ui.setting.component.SettingTitleItem
+import com.mulkkam.ui.setting.model.SettingItem
 import com.mulkkam.ui.setting.model.SettingType
 
 @Composable
