+++ android/app/src/main/java/com/mulkkam/ui/setting/component/SettingNormalItem.kt
@@ -19,7 +19,7 @@ import com.mulkkam.ui.designsystem.Black
 import com.mulkkam.ui.designsystem.Gray400
 import com.mulkkam.ui.designsystem.MulKkamTheme
 import com.mulkkam.ui.designsystem.MulkkamTheme
-import com.mulkkam.ui.setting.SettingItem
+import com.mulkkam.ui.setting.model.SettingItem
 import com.mulkkam.ui.setting.model.SettingType
 
 @Composable
